test:  修改react-native-ble-manager的测试用例
